Standardize on signed integers in quantifiers and loops.
Specifically, use signed "int" for:
1. for loop counters
2. quantified expressions
3. variables and struct fields that count or index into arrays
4. formal parameters for array indexes, offsets and counts

Additionally, strengthen pre-conditions on polyvecl_add()
and polyveck_add() to stabilize proof of those functions
for all parameter sets.

+++ dev/x86_64/src/rej_uniform_avx2.c
@@ -36,8 +36,8 @@
  *            frontend to perform the unintuitive padding.
  */
 
-unsigned int mld_rej_uniform_avx2(
-    int32_t *MLD_RESTRICT r, const uint8_t buf[MLD_AVX2_REJ_UNIFORM_BUFLEN])
+int32_t mld_rej_uniform_avx2(int32_t *MLD_RESTRICT r,
+                             const uint8_t buf[MLD_AVX2_REJ_UNIFORM_BUFLEN])
 {
   unsigned int ctr, pos;
   uint32_t good;
@@ -116,7 +116,7 @@ unsigned int mld_rej_uniform_avx2(
     }
   }
 
-  return ctr;
+  return (int32_t)ctr;
 }
 
 #else /* MLD_ARITH_BACKEND_X86_64_DEFAULT && !MLD_CONFIG_MULTILEVEL_NO_SHARED \
